First attempt to close issue #5

+++ database/ssmw/ssmw.py
@@ -6,6 +6,13 @@ import sys
 import tempfile
 import subprocess
 
+OK = 0
+PARAM_ERR = 78
+DB_AUTH_ERR = 77
+DB_NOT_EXISTS = 66
+MISC_DB_ERR = 70
+WRONG_CREATE_MIGRATE_DIR = 72
+CANNOT_FIND_MIGRATION_PATH = 67
 
 def _bfs_paths(graph, start, goal):
     """
@@ -87,58 +94,63 @@ def shortest_path(creation_path, migration_path, start, goal):
     """
     graph = _build_graph(creation_path, migration_path)
     if start == goal:
-        log.info("Scheme version is %s already, skipping\n" % (start))
         return
     for path in _bfs_paths(graph, start, goal):
-        log.info("Yielding list: %s\n" % (path))
         file_list = _convert2filelist(graph, path)
         yield file_list
-    log.error("There is no any other " +
-                     "possible paths from %s to %s!\n" % (start, goal))
+    raise PathException("There is no any other " +
+                        "possible paths from %s to %s!\n" % (start, goal))
 
 
 class Psql:
-    def __init__(self, db_exec = "psql", db_pass=None, params = ""):
+    def __init__(self, db_exec = "psql", db_pass = None, params = None):
         self.password = db_pass
         self.params   = self._string_param_to_list(params)
         self.psql     = db_exec
 
     def _string_param_to_list(self, string_param):
-        return string_param.split(" ") if string_param != "" else []
+        return string_param.split(" ") if string_param != "" and string_param is not None else []
 
     def execute_query(self, query):
-        try:
-            call = [self.psql] + query + self.params
-            log.info("Executing command %s"%call)
-            stdout = subprocess.check_output(call,
-                                             env=None if db_pass is None else {"PGPASSWORD":self.password})
-            return 0, stdout
-        except subprocess.CalledProcessError as e:
-            return e.returncode, e.output
+        call = [self.psql] + query + self.params
+        log.debug("Executing command %s"%call)
+        p = subprocess.Popen(call,
+                             env=None if db_pass is None else {"PGPASSWORD":self.password},
+                             universal_newlines = True,
+                             stdout=subprocess.PIPE,
+                             stderr=subprocess.PIPE)
+        stdout, stderr = p.communicate()
+        returncode = p.returncode
+
+        if returncode != 0:
+            raise DBException(stderr, returncode)
+        return stdout
 
     def get_db_version(self, db_dbname):
-        errcode, output = self.execute_query(
-            ["-d", db_dbname, "-qc", "copy (select max(id) from version) to stdout"])
-
-        if errcode == 0:
+        try:
+            output = self.execute_query(["-d", db_dbname, "-qc", "copy (select max(id) from version) to stdout"])
             return output
-        else:
-            log.error("There is no db scheme at the moment")
-            return None
+        except DBException as exc:
+            if exc.returncode == 1 and 'ERROR:  relation "version" does not exist' in exc.message:
+                return None
+            # Unknown exception, rethrow
+            raise
 
     def is_db_exists(self, db_dbname):
-        errcode, output = self.execute_query(["-lqtA"])
-        if errcode != 0:
-            # TODO Throw some error
-            return False
-
+        output = self.execute_query(["-lqtA"])
         for line in output.splitlines():
             dbname_ = line.split("|")[0] # Get table name exactly
             if dbname_ == db_dbname:
                 return True
         return False
 
+class DBException(Exception):
+    def __init__(self, message, returncode):
+        super(DBException, self).__init__(message)
+        self.returncode = returncode
 
+class PathException(Exception):
+    pass
 
 # The actual code starts here
 if __name__ == "__main__":
@@ -154,27 +166,59 @@ if __name__ == "__main__":
         db_required_scheme = os.environ["DB_REQUIRED_SCHEME"]
     except KeyError as e:
         log.error("You haven't provided %s environment variable."%e.args[0])
-        sys.exit(78)
+        sys.exit(PARAM_ERR)
 
+    log.debug("""Script was started with following parameters:
+                DB_PARAMS: %s
+                DB_PASS: %s
+                DB_EXEC: %s
+                DB_DBNAME: %s
+                DB_CREATE: %s
+                DB_MIGRATE: %s
+                DB_REQUIRED_SCHEME: %s
+                """%(db_params, db_pass, db_exec, db_dbname, db_create, db_migrate, db_required_scheme)
+              )
 
     instance = Psql(db_exec, db_pass, db_params)
 
     current = None
-    if instance.is_db_exists(db_dbname):
-        current = instance.get_db_version(db_dbname)
-    for filelist in shortest_path(db_create, db_migrate, current, db_required_scheme):
-        fd, merged_sql = tempfile.mkstemp(text=True, suffix=".sql")
-        log.info("Merging %s list of files into file %s"%(filelist, merged_sql))
-        tmpf = os.fdopen(fd, 'w')
-        for file in filelist:
-            with open(file) as f:
-                tmpf.write("\n -- %s\n"%file)
-                tmpf.writelines(f.readlines())
-        tmpf.write("\n")
-        tmpf.close()
-        log.info("Executing file %s as single transaction", merged_sql)
-        err, output = instance.execute_query(["--single-transaction", "-f", merged_sql])
-        if err == 0 and instance.get_db_version(db_dbname) == db_required_scheme:
-            log.info("migration completed!")
-            os.remove(merged_sql)
-            break
+    try:
+        if instance.is_db_exists(db_dbname):
+            current = instance.get_db_version(db_dbname)
+            if current == None:
+                log.info("There is no db version at all")
+        else:
+            log.error("There is no db at the moment")
+            sys.exit(DB_NOT_EXISTS)
+
+        if not (os.path.isdir(db_create) and os.path.isdir(db_migrate)):
+            log.error("Wrong creation or migration path")
+            sys.exit(WRONG_CREATE_MIGRATE_DIR)
+
+        for filelist in shortest_path(db_create, db_migrate, current, db_required_scheme):
+            fd, merged_sql = tempfile.mkstemp(text=True, suffix=".sql")
+            log.info("Merging %s list of files into file %s"%(filelist, merged_sql))
+            tmpf = os.fdopen(fd, 'w')
+            for file in filelist:
+                with open(file) as f:
+                    tmpf.write("\n -- %s\n"%file)
+                    tmpf.writelines(f.readlines())
+            tmpf.write("\n")
+            tmpf.close()
+            log.info("Executing file %s as single transaction", merged_sql)
+            err, output = instance.execute_query(["--single-transaction", "-f", merged_sql])
+            if err == 0 and instance.get_db_version(db_dbname) == db_required_scheme:
+                log.info("migration completed!")
+                os.remove(merged_sql)
+                break
+        log.info("Scheme version is %s already, skipping\n" % current)
+        sys.exit(OK)
+    except DBException as exc:
+        log.error(exc.message)
+        if exc.returncode == 2 and 'FATAL:  password authentication failed for user' in exc.message:
+            sys.exit(DB_AUTH_ERR)
+        else:
+            sys.exit(MISC_DB_ERR)
+    except PathException as exc:
+        log.error(exc.message)
+        sys.exit(CANNOT_FIND_MIGRATION_PATH)
